- Moved LoadApiDataFacade to the api_infrastructure folder
